Move AreenaPlaylistParser to a dedicated file
